feat: add docker configuration with MongoDB

+++ notes.controller.js
@@ -1,7 +1,7 @@
 const Note = require("./models/Note");
 
-async function addNote(title, { id } = {}) {
-  const note = await Note.create({ title });
+async function addNote(title, owner) {
+  const note = await Note.create({ title, owner });
   return note;
 }
 
@@ -10,15 +10,17 @@ async function getNotes() {
   return notes;
 }
 
-async function deleteNoteById(id) {
-  const res = await Note.deleteOne({ _id: id });
+async function deleteNoteById(id, owner) {
+  const res = await Note.deleteOne({ _id: id, owner });
+  if (res.deletedCount === 0) throw new Error("Note not found");
   return res.deletedCount;
 }
-async function updateNote(id, updatedData) {
+async function updateNote(id, updatedData, owner) {
   const res = await Note.findOneAndUpdate(
-    { _id: id },
+    { _id: id, owner },
     { title: updatedData.title }
   );
+  if (res === null) throw new Error("Note not found");
   return res;
 }
